+++ 최규화/Mod_Server/Server/PhysicsEngine/MyMiniPysicsEngine.cpp
@@ -438,7 +438,7 @@ bool MiniPhysicsEngineG9::PhysicsPoint::CollisionTest(PhysicsPoint & p2, XMFLOAT
 	return false;
 }
 
-void MiniPhysicsEngineG9::PhysicsPoint::CollisionResolve(PhysicsPoint & p2, XMFLOAT3 & CollisionN, __int64 DeltaTime)
+void MiniPhysicsEngineG9::PhysicsPoint::CollisionResolve(PhysicsPoint & p2, XMFLOAT3 & CollisionN, double DeltaTime)
 {
 	//충돌후 속도를 계산함.
 	ResolveVelocity(p2, CollisionN, DeltaTime);
@@ -462,7 +462,7 @@ bool MiniPhysicsEngineG9::PhysicsPoint::GetBounce()
 	return Bounce;
 }
 
-void MiniPhysicsEngineG9::PhysicsPoint::ResolveVelocity(PhysicsPoint & p2, XMFLOAT3 & CollisionN, __int64 DeltaTime)
+void MiniPhysicsEngineG9::PhysicsPoint::ResolveVelocity(PhysicsPoint & p2, XMFLOAT3 & CollisionN, double DeltaTime)
 {
 	//충돌후 속도 변화는 다음과 같다.
 	//1. 충돌전 분리속도를 구하고
@@ -537,7 +537,7 @@ void MiniPhysicsEngineG9::PhysicsPoint::ResolveVelocity(PhysicsPoint & p2, XMFLO
 //만약 이게 마음에 안들면 x와 z의 속도가 0이면 밀려나지 않도록 하는 방법이 있음.
 //물론 이게 또 문제가 뭐냐면 아예 완전히 겹쳐있는 순간에 안움직이고 있으면 안떼어내진다.
 //나는 그냥 공중에서는 밀어내도 상관 없도록 하는게 나을것 같음.
-void MiniPhysicsEngineG9::PhysicsPoint::ResolvePenetration(PhysicsPoint & p2, __int64 DeltaTime)
+void MiniPhysicsEngineG9::PhysicsPoint::ResolvePenetration(PhysicsPoint & p2, double DeltaTime)
 {
 	if (penetration < MMPE_EPSILON)//밀어낼 필요가 없는경우
 		return;
